Remove unnecessary code
This removes the same unnecessary code as in 2d44627 but it also
selects the first element in the array that's returned from calling
the `split` method.

We could also use square braces to select the first element like so:

`movie_name: tr.at('.titleColumn a').text.split(/"/)[0],`

But using `.first` is more human readable so we prefer that.